Broad refactor and docstrings
