+++ PythonRegression/util/static_vals.py
@@ -18,6 +18,11 @@ TEST_CONSISTENCY_ADDRESS = "THIS9TRANSACTION9IS9NOT9CONSISTENT999999999999999999
 TEST_BLOWBALL_COO = "EFPNKGPCBXXXLIBYFGIGYBYTFFPIOQVNNVVWTTIYZO9NFREQGVGDQQHUUQ9CLWAEMXVDFSSMOTGAHVIBH"
 
 
+SIDE_TANGLE_ADDRESS = "SIDE9TANGLE9999999999999999999999999999999999999999999999999999999999999999999999"
+STITCHING_ADDRESS = "STITCHING9TRANSACTIONS99999999999999999999999999999999999999999999999999999999999"
+REFERENCING_ADDRESS = "REFERENCES9STITCHING9TRANSACTION9999999999999999999999999999999999999999999999999"
+
+
 SIDE_TANGLE_TRANSACTIONS =[
     'MKSVLU9BBBKOIAMPHKCHWNKXESH9RWOOAPSJUZXPDIQIMBLIEXLNARBNPKNCHPEZSYBXWUVEKVJMXX999',
     'HADZDR9WDCRFTNRPXZTUUSZYXPRHIGVANQQMHVPGIBDDKEFCWVPAXNZUNEKDAOIOYTVPKOLRNNGVSE999',
